+++ filament/include/filament/Skybox.h
@@ -23,8 +23,9 @@
 
 #include <utils/compiler.h>
 
-#include <stdint.h>
 #include <math/mathfwd.h>
+
+#include <stdint.h>
 
 namespace filament {
 
@@ -103,7 +104,7 @@ public:
          *
          * @see Texture
          */
-        Builder& environment(Texture* cubemap) noexcept;
+        Builder& environment(Texture* UTILS_NONNULL cubemap) noexcept;
 
         /**
          * Indicates whether the sun should be rendered. The sun can only be
@@ -178,9 +179,9 @@ public:
          *
          * @param engine Reference to the filament::Engine to associate this Skybox with.
          *
-         * @return pointer to the newly created object, or nullptr if the light couldn't be created.
+         * @return pointer to the newly created object.
          */
-        Skybox* build(Engine& engine);
+        Skybox* UTILS_NONNULL build(Engine& engine);
 
     private:
         friend class FSkybox;
@@ -229,7 +230,7 @@ public:
     /**
      * @return the associated texture, or null if it does not exist
      */
-    Texture const* getTexture() const noexcept;
+    Texture const* UTILS_NONNULL getTexture() const noexcept;
 
     /**
      * Changes the coarse-level draw ordering.